Add a slash command for reloading Mappy
Add a slash command for reloading the current Mappy profile.
This is meant to fix small issues in case the buttons didn't
render in time for the Mappy stacking to properly position them.

If the buttons overlap suddenly after a login screen, just
type /mappy reload to fix them. It's a rare occurence, but can
happen.